Add input-ui-state
The input-ui-state is responsible for keeping track of single/double mouse
clicks and such. This could also be part of the ui-state, but the idea is to
make this input-ui-state not application specific. Maybe in the future it
will be changed to a library that can be used by other applications as well.