Let double mouse click (also) toggle cell
A single mouse click has a delay of 500 ms because it needs to wait to see if
there is a second mouse click (in which case it is actually a double mouse
click). This delay feels a little awkward. For people that prefer, a double
mouse click will now also toggle the cell state.