Treat cell as just [x y] coordinates
The functions board/make-cell and board/cell-coords have been introduced to
create a cell abstraction. The idea was that in the future the cell might be
something more than just its coordinates, but I don't see this happening. So I
decided to simplify the code and treat a cell as just a vector of x and y.